# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Beauty Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sms/settings:
    get:
      operationId: getSmsSettings
      tags: [sms]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsSettings"
    put:
      operationId: updateSmsSettings
      tags: [sms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsSettingsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsSettings"
  /sms/templates:
    get:
      operationId: getSmsTemplates
      tags: [sms]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsTemplates"
    put:
      operationId: updateSmsTemplates
      tags: [sms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsTemplatesInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsTemplates"
...
  /sms/send:
    post:
      operationId: sendManualSms
      tags: [sms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualSmsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualSmsResult"
  /sms/send-pattern:
    post:
      operationId: sendPatternSms
      tags: [sms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatternSmsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatternSmsResult"
  /sms/saved-patterns:
    get:
      operationId: listSavedPatterns
      tags: [sms]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedPattern"
    post:
      operationId: createSavedPattern
      tags: [sms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedPatternInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPattern"
  /sms/saved-patterns/{id}:
    delete:
      operationId: deleteSavedPattern
      tags: [sms]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /sms/logs:
    get:
      operationId: listSmsLogs
      tags: [sms]
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsLogList"

  # ── Waiting list ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SmsSettings:
      type: object
      required: [username, from, hasPassword, enabledAppointment, enabledPayment, enabledCommission, enabledRecipientWelcome, enabledSurvey, surveyThrottleDays, sendMode, bodyIdAppointment, bodyIdPayment, bodyIdCommission, bodyIdBirthday, bodyIdSurvey, bodyIdRecipientWelcome]
      properties:
        username: { type: string }
        from: { type: string }
        hasPassword: { type: boolean }
        enabledAppointment: { type: boolean }
        enabledPayment: { type: boolean }
        enabledCommission: { type: boolean }
        enabledRecipientWelcome: { type: boolean }
        enabledSurvey: { type: boolean }
        surveyThrottleDays: { type: integer }
        sendMode: { type: string, enum: [normal, pattern] }
        bodyIdAppointment: { type: string }
        bodyIdPayment: { type: string }
        bodyIdCommission: { type: string }
        bodyIdBirthday: { type: string }
        bodyIdSurvey: { type: string }
        bodyIdRecipientWelcome: { type: string }
    SmsSettingsInput:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
        from: { type: string }
        enabledAppointment: { type: boolean }
        enabledPayment: { type: boolean }
        enabledCommission: { type: boolean }
        enabledRecipientWelcome: { type: boolean }
        enabledSurvey: { type: boolean }
        surveyThrottleDays: { type: integer }
        sendMode: { type: string, enum: [normal, pattern] }
        bodyIdAppointment: { type: string }
        bodyIdPayment: { type: string }
        bodyIdCommission: { type: string }
        bodyIdBirthday: { type: string }
        bodyIdSurvey: { type: string }
        bodyIdRecipientWelcome: { type: string }
    SmsTemplates:
      type: object
      required: [appointment, payment, commission, birthday, survey, recipientWelcome, defaults]
      properties:
        appointment: { type: string }
        payment: { type: string }
        commission: { type: string }
        birthday: { type: string }
        survey: { type: string }
        recipientWelcome: { type: string }
        defaults:
          type: object
          required: [appointment, payment, commission, birthday, survey, recipientWelcome]
          properties:
            appointment: { type: string }
            payment: { type: string }
            commission: { type: string }
            birthday: { type: string }
            survey: { type: string }
            recipientWelcome: { type: string }
    SmsTemplatesInput:
      type: object
      properties:
        appointment: { type: string }
        payment: { type: string }
        commission: { type: string }
        birthday: { type: string }
        survey: { type: string }
        recipientWelcome: { type: string }
...
    ManualSmsInput:
      type: object
      required: [message]
      properties:
        message: { type: string }
        patientIds:
          type: array
          items: { type: integer }
        birthdayDays: { type: integer }
        eventType: { type: string }
    ManualSmsResult:
      type: object
      required: [total, sent, failed]
      properties:
        total: { type: integer }
        sent: { type: integer }
        failed: { type: integer }
        errors:
          type: array
          items: { type: string }
    PatternSmsInput:
      type: object
      required: [bodyId]
      properties:
        bodyId: { type: string }
        args:
          type: array
          items: { type: string }
        patientId: { type: integer }
        phone: { type: string }
    PatternSmsResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        error: { type: ["string", "null"] }
    SavedPattern:
      type: object
      required: [id, name, bodyId]
      properties:
        id: { type: integer }
        name: { type: string }
        bodyId: { type: string }
    SavedPatternInput:
      type: object
      required: [name, bodyId]
      properties:
        name: { type: string }
        bodyId: { type: string }
    SmsLog:
      type: object
      required: [id, recipientPhone, eventType, message, status, createdAt]
      properties:
        id: { type: integer }
        recipientPhone: { type: string }
        recipientName: { type: ["string", "null"] }
        patientId: { type: ["integer", "null"] }
        eventType: { type: string }
        message: { type: string }
        status: { type: string }
        error: { type: ["string", "null"] }
        createdAt: { type: integer }
    SmsLogList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SmsLog"
        total: { type: integer }

    # ── Waiting list ──────────────────────────────────────────────────────────
...